requirements.txt
Django==3.0.7
django-allauth==0.42.0
django-countries==6.1.2
django-crispy-forms==1.9.1
Pillow==7.1.1
requests
stripe==2.48.0
python-decouple==3.3
sslcommerz-python==0.0.7
mysqlclient
selenium
pandas
mysql-connector
django-paypal